Update README.md
/* You will be creating a TicTacToe game, I have provided a shell of a program that you will complete.  You are required to write 5 functions but more than likely you will have to write numerous private helper functions.  The functions you have to write are, you are not allowed to change any function signatures or remove any of my code:

isValidMove - This function says if there the given position is currently occupied
printBoard - This function will print the TicTacToe board
hasWon - This function takes a player and returns true if they won
isTie - Returns true if the game is a tie, false if not, this works only if all spots on the board are occupied.  False is returned if not all spots are occupied
markTheBoard - This function is purely to help with the test program that will test this class.  It allows me to arbitrarily mark the board, even marking as empty to help clear the board.  This function should not do any validation, it should just place the given marking at the given position.*/